requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.3
cached-property==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2
django-admin==2.0.1
django-ckeditor==5.9.0
django-elasticsearch-dsl==7.1.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filer==1.7.0
django-jalali-date-3==0.2.1
django-js-asset==1.2.2
django-json-widget==1.0.1
django-mptt==0.11.0
django-nine==0.2.3
django-polymorphic==2.1.2
django-redis==4.11.0
django-silk==4.0.1
django-six==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
drf-yasg==1.17.1
easy-thumbnails==2.7
elasticsearch==7.6.0
elasticsearch-dsl==7.2.0
fcm-django==0.3.4
future==0.18.2
gprof2dot==2019.11.30
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
isodate==0.6.0
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
kavenegar==1.1.2
lxml==4.5.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.3
Pillow==7.1.1
pycodestyle==2.6.0
pyfcm==1.4.7
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.1
requests==2.23.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
screen==1.0.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
xlwt==1.3.0
zeep==3.4.0
zipp==3.1.0

django-rest-auth~=0.9.5